Create appropriate ImageView per CardView
For each CardView in RecyclerView god_recycler_view, create an ImageView which should correspond to the god or goddess being represented. An image attribute has been added to the God data class and this attribute has been accounted for in GodAdapter and GodViewHolder. All PNG images have been included as drawable resources.

When an image asset cannot be found as a drawable resource, drawable gnome_child is set as the fallback value.